news_search_summarized/requirements.txt
streamlit

langchain
langchain-core
langchain-community
langchain-groq
langgraph

pydantic
python-dotenv

Tavily

groq